did... the project
I wasn't thinking originally of uploading this to git, but in the end I didn't want to spare myself the bragging rights (as if!). Anyway this lame message is meant convey that I shoved all the code into one commit in a semi hilarious way no one will probably see.